Updated failing execution tests: semanticTests/{constructor, functionTypes, payable}.

+++ test/libsolidity/semanticTests/functionTypes/function_type_library_internal.sol
@@ -1,23 +1,23 @@
 library Utils {
     function reduce(
-        uint256[] memory array,
+        uint[] memory array,
         function(uint, uint) internal returns (uint) f,
-        uint256 init
-    ) internal returns (uint256) {
-        for (uint256 i = 0; i < array.length; i++) {
+        uint init
+    ) internal returns (uint) {
+        for (uint i = 0; i < array.length; i++) {
             init = f(array[i], init);
         }
         return init;
     }
 
-    function sum(uint256 a, uint256 b) internal returns (uint256) {
+    function sum(uint a, uint b) internal returns (uint) {
         return a + b;
     }
 }
 
 
 contract C {
-    function f(uint256[] memory x) public returns (uint256) {
+    function f(uint[] memory x) public returns (uint) {
         return Utils.reduce(x, Utils.sum, 0);
     }
 }
@@ -25,4 +25,4 @@ contract C {
 // ====
 // compileViaYul: also
 // ----
-// f(uint256[]): 0x20, 0x3, 0x1, 0x7, 0x3 -> 11
+// f(uint[]): dynarray 0 [1, 7, 3] -> 11
